Move CurrencyToString to size
